Fix sam_knn.py variable name for reset
As presented here, https://github.com/scikit-multiflow/scikit-multiflow/issues/272, `SAMKNNClassifier` cannot use the `reset()` method. This seems to be due to the name mismatch between internal variable `self.max_wind_size` and input argument `max_window_size`.

`max_wind_size` does not appear to be referenced anywhere else in the library, so simply editing the name to match `max_window_size` in `SAMKNNClassifier.__init__()` may be all the change that is needed to fix this.